requirements.txt
# Core Data Science Libraries
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.3
# Machine Learning & NLP
transformers==4.35.0
sentence-transformers==2.2.2
torch==2.1.0
# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.21
openpyxl==3.1.2
# Text Processing
spacy==3.7.2
nltk==3.8.1
regex==2023.10.3
# Web Framework & API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
# Caching & Queue
redis==5.0.1
celery==5.3.4
# Frontend
streamlit==1.28.0
plotly==5.17.0
matplotlib==3.8.0
seaborn==0.13.0
# MLOps & Monitoring
mlflow==2.8.0
prometheus-client==0.19.0
# Vector Database
pinecone-client==2.2.4
# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
tqdm==4.66.1
# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.1
# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.0
